fix(docs): re-ordered small paragraphs, clarified wording, and added links to tech homepages (#531)
# Summary

As a new Unicorn, I made changes which I believe would've increased my
comprehension of this document on the first read-through.

## Description

- Re-ordered tools in monitoring to build conceptually in line with the
way each enables the next
- Turned first mention of each tool into a link to the relevant docs.
Both an ease of reference change and also to disambiguate in the event a
name collision exists in the reader's mind.
- Clarified relationship between AuthService and Keycloak as this
Unicorn initially saw a redundancy and suspected AuthService was perhaps
the service IAM solution while Keycloak was the user IAM solution. We
could link out to BigBang's explanation of this relationship but I did
not:
https://docs-bigbang.dso.mil/2.2.0/docs/understanding-bigbang/package-architecture/authservice/
- Wording changes for precision. Example, in MetricsServer, speaking of
"container" metrics is more precise (and I believe more accurate) than
"cluster" metrics. If it reports on nodes and containers I would
(personally) speak of both over the more nebulous "cluster" (which has
many more primitives (notably servics and pods) which could be but are
not (I think) being reported on).
- Most likely errors needing correction:
  - Metrics Server may be misunderstood
- AuthService may be better characterized but I didn't want to plagarize
the BigBang docs, and am new to this service.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)(https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md#submitting-a-pull-request)
followed

---------

Co-authored-by: Micah Nagel <micah.nagel@defenseunicorns.com>