docs: add documentation on metrics/dashboards for apps (#1221)
## Description

Adds/modifies a docs page detailing how metrics can be collected for an
additional application (outside core) and how to add on additional
dashboards.

This doc is partially being added in response to a support thread
requesting info on how to group dashboards within uds core when adding
new ones on. Related to that grouping setup, I also added support for
annotations on the loki dashboard configmap.

I also moved some of the information around our monitoring setup and why
we made the choices we did into a new dev doc for this purpose. While it
still has valuable information, it's not as important for the end user
to know/understand unless they wanted to dig into why we chose our
approach (in which case the dev docs are a good place for them to be
looking).

## Related Issue

N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Steps to Validate

If you want to test the grouping, that can be done by making a new
bundle (or modifying the standard bundle locally) and confirming that
all the core dashboards end up grouped together.

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed