chore: remove watch and conditional logic around ambient component (#1447)
## Description

As a result of https://github.com/defenseunicorns/uds-core/pull/1428, we
are now guaranteed that the ambient component is deployed with the base
layer. The PR removes code that handled "conditional ambient components"
with a fallback to sidecar mode if the component was not deployed.

Note: If we did require a revert of the required ambient component for
any reason, this PR should be reverted as well (to re-add the handling
of conditional ambient).

## Related Issue

N/A

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [c] Other (security config, docs update, etc)

## Steps to Validate

Tests should adequately cover this change (noting that the fallback test
was deleted).

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed