feat: implement ambient authservice applications (#1716)
## Description
Implementation of Ambient enabled authservice applications. This means
the operator will manage the creation of waypoint related resources so
that the L7 layer protection is intact and routable. The design doc for
this implementation can be found
[here](https://www.notion.so/defense-unicorns/Ambient-Authservice-Applications-214e512f24fc80b3a53bd2d3bcef80e8).

## Related Issue

Fixes #1660 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate
- `uds run test-uds-core`
- deploying additional custom applications that are ambient and
authservice enabled on top of that cluster

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed