fix(ci): snapshot release creation fix (#1786)
## Description

The snapshot release has been broken since the introduction of the
base-shim bundle -
https://github.com/defenseunicorns/uds-core/actions/workflows/snapshot-release.yaml

This PR makes the base-shim bundle optional with a new input, since it
shouldn't be created when we are making the base layer for the purpose
of other bundles (i.e. slim-dev).

## Related Issue

N/A

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Steps to Validate

Validated by running snapshot on this branch:
https://github.com/defenseunicorns/uds-core/actions/runs/16724528903

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed