fix: replace neuvector in diagrams with falco (#1964)
## Description
Updating the drawio diagrams to have falco instead of neuvector. Much of
this will probably be redone or recreated with C4 in coming weeks so not
spending a ton of time to make this perfect.

Because these are stored in git, we can reference neuvector diagrams
historically if necessary. Google drive has also been updated with the
master drawio files.

## Related Issue

Fixes #1916

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor
Guide](https://github.com/defenseunicorns/uds-template-capability/blob/main/CONTRIBUTING.md)
followed