chore(ci): handle timing with pepr joining the mesh (#2040)
## Description

Adjusts the order such that Pepr is pre-labelled for ambient mode before
Istio is installed. This seems to adjust the timing in testing such that
ztunnel is ready to accept connections on Pepr's behalf and we do not
see a race condition in CI.

Also cleans up some sidecar upgrade actions that shouldn't be necessary
anymore.

## Related Issue

Related to intermittent CI issues.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)